chore: move integration test into `PeerGradingServiceIntegrationTest` class
